DSE: Merge Flag/GPR/FPR passes for perf
