+++ External/FEXCore/Source/CMakeLists.txt
@@ -155,10 +155,8 @@ set (SRCS
   Interface/IR/Passes/ValueDominanceValidation.cpp
   Interface/IR/Passes/PhiValidation.cpp
   Interface/IR/Passes/RedundantFlagCalculationElimination.cpp
-  Interface/IR/Passes/DeadFlagStoreElimination.cpp
+  Interface/IR/Passes/DeadStoreElimination.cpp
   Interface/IR/Passes/StaticRegisterAllocationPass.cpp
-  Interface/IR/Passes/DeadGPRStoreElimination.cpp
-  Interface/IR/Passes/DeadFPRStoreElimination.cpp
   Interface/IR/Passes/RegisterAllocationPass.cpp
   Interface/IR/Passes/SyscallOptimization.cpp
   Utils/ELFLoader.cpp
