+++ External/FEXCore/Source/Interface/IR/Passes.h
@@ -8,9 +8,7 @@ FEXCore::IR::Pass* CreateConstProp(bool InlineConstants);
 FEXCore::IR::Pass* CreateContextLoadStoreElimination();
 FEXCore::IR::Pass* CreateSyscallOptimization();
 FEXCore::IR::Pass* CreateDeadFlagCalculationEliminination();
-FEXCore::IR::Pass* CreateDeadFlagStoreElimination();
-FEXCore::IR::Pass* CreateDeadGPRStoreElimination();
-FEXCore::IR::Pass* CreateDeadFPRStoreElimination();
+FEXCore::IR::Pass* CreateDeadStoreElimination();
 FEXCore::IR::Pass* CreatePassDeadCodeElimination();
 FEXCore::IR::Pass* CreateIRCompaction();
 FEXCore::IR::RegisterAllocationPass* CreateRegisterAllocationPass(FEXCore::IR::Pass* CompactionPass, bool OptimizeSRA);
